introduce RawAvroUtils
